add multi-language loading skeleton
